requirements.txt
torch
numpy
pygame
matplotlib
Pillow